Fixed CR comments
